Allow reading from yaml config
Use String instead of "string slice" &str in Config struct
Why? String is heap-allocated, so using this

```rust
pub struct Config {
   dirs: Vec<String>
}
```
... lets us avoid lifetime headaches.

Also moved main.rs to /bin subdirectory. This is a cargo convention that
aids the rust compiler. Similarly, modules intended for export need to
be listed in lib.rs

Note that all tests that run under /tests are compiled as a separate
crate, and should be thought of as integration tests. Unit tests can be
defined in library files themselves.